Moved commom EntityControllerCacheTest functions to trait.
